PeakSeeking/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "peak-seeking"
version = "0.1.0"
description = "Peak-seeking RL demo"
authors = [{ name = "Your Name" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "gymnasium>=0.29",
    "numpy>=1.24",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
where = ["."]